+++ modules/codemod/lib/v14-tokens/spec/migrateColorTokens.spec.ts
@@ -84,8 +84,8 @@ describe('Canvas Kit Tokens > Canvas Tokens v3', () => {
         `;
 
       const expected = stripIndent`
-          import { base } from "@workday/canvas-tokens-web";
           import { cssVar } from "@workday/canvas-kit-styling";
+          import { base } from "@workday/canvas-tokens-web";
 
           const color = cssVar(base.blue600);
         `;
@@ -120,8 +120,8 @@ describe('Canvas Kit Tokens > Canvas Tokens v3', () => {
         `;
 
       const expected = stripIndent`
-          import { base } from "@workday/canvas-tokens-web";
           import { cssVar } from "@workday/canvas-kit-styling";
+          import { base } from "@workday/canvas-tokens-web";
 
           const color = cssVar(base.blue600);
           const color2 = cssVar(base.red800);
@@ -298,8 +298,8 @@ describe('Canvas Kit Tokens > Canvas Tokens v3', () => {
         `;
 
       const expected = stripIndent`
-          import { cssVar } from "@workday/canvas-kit-styling";
           import { system } from "@workday/canvas-tokens-web";
+          import { cssVar } from "@workday/canvas-kit-styling";
 
           const styles = css({
             background: cssVar(system.color.bg.default),
@@ -372,8 +372,8 @@ describe('Canvas Kit Tokens > Canvas Tokens v3', () => {
         `;
 
       const expected = stripIndent`
-          import { cssVar } from "@workday/canvas-kit-styling";
           import { system } from "@workday/canvas-tokens-web";
+          import { cssVar } from "@workday/canvas-kit-styling";
 
           const styles = {
             color: toggled ? cssVar(system.color.fg.primary.default) : cssVar(system.color.fg.default),